Restructure for code reuse and hiding implementation.
